ENH: Major SOFB refactoring, bug fix and functionality implementation.
