+++ pyqt-apps/siriushla/as_ap_sofb/__init__.py
@@ -1,0 +1,1 @@
+from siriushla.as_ap_sofb.main import MainWindow
