--- pyqt-apps/siriushla/si_ap_sofb/__init__.py
@@ -1,3 +1,0 @@
-from siriushla.si_ap_sofb.si_ap_sofb import *
-
-__all__ = ('si_ap_sofb',)
